requirements.txt
tensorflow
numpy
opencv-python
pillow
matplotlib
scikit-learn